Separate code for backward compatibility to a separate file.
